requirements.txt
streamlit
openai-whisper
yt-dlp
pydub
numpy<2
tiktoken
torch
torchvision
torchaudio